With this fix the following scenarios work successfully:
* Restarting Framework
* JN Standalone
* JN Colocated
* DN
* 2 DNS
* Active NN
* Standby NN
* JN + NN Not Colocated
* 2 JNS
* ZKFC (track ZKFC and then test killing and relaunching it)
* JN + NN Colocated
* Both NNS

We still need to test taking the full machine down so nodes cannot
recover on the same machine. I will be testing that with a subsequent
PR. I will also create issues for failure cases which we need to handle
by using a NN backup image (i.e. both machine holding the NN’s go down
and we need to start them on new machines).