Add timeout for dead nodes/ Sym Link Code Update
This fixed the failure scenario for bringing up a node on a new host
instead of waiting forever for a dead node to come up. I also fixed the
sym link code to remove a dangling sym link in the case that a sym link
is left dangling on an empty folder.