Package Java and Hadoop with HDFS
In the testing/continuous branch, frameworks have had their access to
the host's java and hadoop binaries/scripts revoked.  In order for
frameworks to use java and hadoop it must be packaged with the
framework.

This change along with changes made to the mesosphere/universe repo
packages the needed java and hadoop dependencies with the HDFS
framework.