Consistent Configs
I missed a couple of things in reviewing the last PR so I am fixing
them here:

1) Our standard in HdfsFrameworkConfig is that numeric values are
constant, but strings are not. Probably strings should also be
constants, but meanwhile I wanted to make sure we are consistent. We
also might want to consider another class in the future to store these
constants

2) I realized our configurations are consistent. Running on linux,
which is what our open source users do, typically the mesos library
would be under /usr/local/lib and that is our default configuration. In
DCOS we have a different configuration. I have updated that.