Refactor Scheduler
This change makes the Scheduler adopt a more object-oriented style.  The
different node types get their own classes and inherit from a base
class.  That base class requires implementation of a couple interfaces
allowing for a consistent pattern of evaluating and accepting/rejecting
offers.

A significant amount of logic is removed from the Scheduler class itself
and placed in the Node class hierarchy.  Additional refactoring work is
still needed espcially in the HdfsNode class, but this change produces
a significant restructure of the Scheduler.